INVENTORY MANAGEMENT SYSTEM
This is a simple console-based project made in C++ to help manage products in a store or warehouse. It lets you add new products, view all products, search by product ID, update product details, and delete products. All the data is saved in a text file(inventory.txt) so you don’t lose anything when the program is closed.